move footer to dedicated file
